new viewmode setting
